#13067 [feature] metricManualEntity 수정사항 변경

+++ src/main/kotlin/co/brainz/itsm/sla/metricManual/dto/MetricManualKeyDto.kt
@@ -6,5 +6,5 @@ import java.time.LocalDate
 data class MetricManualKeyDto(
     var metricId: String ="",
     var referenceDt: LocalDate? = null,
-    var metricValue: Int = 0
+    var metricValue: Double = 0.0
 ): Serializable
